# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test work 1",
        "description": "Test work of Rest Api by Laravel 11 platform",
        "contact": {
            "name": "ASt",
            "email": "alexandr.statut@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Dev API server"
        },
        {
            "url": "http://127.0.0.1:8001/api",
            "description": "Prod API server"
        }
    ],
    "paths": {
...
        "/v1/image/{id}": {
            "get": {
                "tags": [
                    "Изображения"
                ],
                "summary": "Api: Получение отдельного изображения",
                "description": "Получение отдельного изображения",
                "operationId": "v1.image.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID товара",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ImageResource"
                                        }
                                    },
                                    "type": "object"
                                },
                                "examples": {
                                    "Some File": {
                                        "summary": "Some File",
                                        "value": {
                                            "id": 1,
                                            "path": "avatar",
                                            "filename": "test",
                                            "ext": "png",
                                            "created": "2024-03-01 11:00:00",
                                            "updated": "2024-03-01 11:00:00"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Изображения"
                ],
                "summary": "Api: Удаление изображения",
                "description": "Удаление изображения",
                "operationId": "v1.iamge.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID файла",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MessageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResource": {
                "title": "Ответ с ошибкой",
                "properties": {
                    "status": {
                        "title": "Статус",
                        "type": "boolean",
                        "format": "int64",
                        "example": "false"
                    },
                    "code": {
                        "title": "Код",
                        "type": "integer",
                        "format": "int64",
                        "example": "500"
                    },
                    "msg": {
                        "title": "Сообщение",
                        "type": "string",
                        "example": "Test"
                    }
                },
                "type": "object"
            },
            "ImageResource": {
                "title": "Карточка изображения",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Transform the resource into an array.",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "path": {
                        "title": "Каталог",
                        "type": "string",
                        "example": "avatar"
                    },
                    "filename": {
                        "title": "Имя файла",
                        "type": "string",
                        "example": "test"
                    },
                    "ext": {
                        "title": "Расширение файла",
                        "type": "string",
                        "example": "png"
                    },
                    "url": {
                        "title": "URL для получения",
                        "type": "string",
                        "example": "https://site.com/test.png"
                    },
                    "created": {
                        "title": "Создана",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-01 12:00:00"
                    },
                    "updated": {
                        "title": "Обновлена",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-01 12:00:00"
                    }
                },
                "type": "object"
            },
...
            "MessageResource": {
                "title": "Простой ответ",
                "properties": {
                    "status": {
                        "title": "Статус",
                        "description": "Transform the resource into an array.",
                        "type": "boolean",
                        "example": "true"
                    },
                    "msg": {
                        "title": "Сообщение",
                        "type": "string",
                        "example": "Test"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Your token providing after auth in Api",
                "name": "Api Client",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}